+++ lib/dns.js
@@ -22,11 +22,8 @@
 'use strict';
 
 const {
-  ArrayPrototypeMap,
-  ObjectCreate,
   ObjectDefineProperties,
   ObjectDefineProperty,
-  ReflectApply,
   Symbol,
 } = primordials;
 
@@ -37,15 +34,16 @@ const { customPromisifyArgs } = require('internal/util');
 const errors = require('internal/errors');
 const {
   bindDefaultResolver,
-  getDefaultResolver,
   setDefaultResolver,
-  Resolver,
   validateHints,
   emitInvalidHostnameWarning,
   getDefaultVerbatim,
   setDefaultResultOrder,
   errorCodes: dnsErrorCodes,
 } = require('internal/dns/utils');
+const {
+  Resolver
+} = require('internal/dns/callback_resolver');
 const {
   NODATA,
   FORMERR,
@@ -89,12 +87,10 @@ const {
 const {
   GetAddrInfoReqWrap,
   GetNameInfoReqWrap,
-  QueryReqWrap,
 } = cares;
 
 const kPerfHooksDnsLookupContext = Symbol('kPerfHooksDnsLookupContext');
 const kPerfHooksDnsLookupServiceContext = Symbol('kPerfHooksDnsLookupServiceContext');
-const kPerfHooksDnsLookupResolveContext = Symbol('kPerfHooksDnsLookupResolveContext');
 
 const {
   hasObserver,
@@ -293,91 +289,6 @@ function lookupService(address, port, callback) {
 ObjectDefineProperty(lookupService, customPromisifyArgs,
                      { __proto__: null, value: ['hostname', 'service'], enumerable: false });
 
-
-function onresolve(err, result, ttls) {
-  if (ttls && this.ttl)
-    result = ArrayPrototypeMap(
-      result, (address, index) => ({ address, ttl: ttls[index] }));
-
-  if (err)
-    this.callback(dnsException(err, this.bindingName, this.hostname));
-  else {
-    this.callback(null, result);
-    if (this[kPerfHooksDnsLookupResolveContext] && hasObserver('dns')) {
-      stopPerf(this, kPerfHooksDnsLookupResolveContext, { detail: { result } });
-    }
-  }
-}
-
-function resolver(bindingName) {
-  function query(name, /* options, */ callback) {
-    let options;
-    if (arguments.length > 2) {
-      options = callback;
-      callback = arguments[2];
-    }
-
-    validateString(name, 'name');
-    validateFunction(callback, 'callback');
-
-    const req = new QueryReqWrap();
-    req.bindingName = bindingName;
-    req.callback = callback;
-    req.hostname = name;
-    req.oncomplete = onresolve;
-    req.ttl = !!(options && options.ttl);
-    const err = this._handle[bindingName](req, toASCII(name));
-    if (err) throw dnsException(err, bindingName, name);
-    if (hasObserver('dns')) {
-      startPerf(req, kPerfHooksDnsLookupResolveContext, {
-        type: 'dns',
-        name: bindingName,
-        detail: {
-          host: name,
-          ttl: req.ttl,
-        },
-      });
-    }
-    return req;
-  }
-  ObjectDefineProperty(query, 'name', { __proto__: null, value: bindingName });
-  return query;
-}
-
-const resolveMap = ObjectCreate(null);
-Resolver.prototype.resolveAny = resolveMap.ANY = resolver('queryAny');
-Resolver.prototype.resolve4 = resolveMap.A = resolver('queryA');
-Resolver.prototype.resolve6 = resolveMap.AAAA = resolver('queryAaaa');
-Resolver.prototype.resolveCaa = resolveMap.CAA = resolver('queryCaa');
-Resolver.prototype.resolveCname = resolveMap.CNAME = resolver('queryCname');
-Resolver.prototype.resolveMx = resolveMap.MX = resolver('queryMx');
-Resolver.prototype.resolveNs = resolveMap.NS = resolver('queryNs');
-Resolver.prototype.resolveTxt = resolveMap.TXT = resolver('queryTxt');
-Resolver.prototype.resolveSrv = resolveMap.SRV = resolver('querySrv');
-Resolver.prototype.resolvePtr = resolveMap.PTR = resolver('queryPtr');
-Resolver.prototype.resolveNaptr = resolveMap.NAPTR = resolver('queryNaptr');
-Resolver.prototype.resolveSoa = resolveMap.SOA = resolver('querySoa');
-Resolver.prototype.reverse = resolver('getHostByAddr');
-
-Resolver.prototype.resolve = resolve;
-
-function resolve(hostname, rrtype, callback) {
-  let resolver;
-  if (typeof rrtype === 'string') {
-    resolver = resolveMap[rrtype];
-  } else if (typeof rrtype === 'function') {
-    resolver = resolveMap.A;
-    callback = rrtype;
-  } else {
-    throw new ERR_INVALID_ARG_TYPE('rrtype', 'string', rrtype);
-  }
-
-  if (typeof resolver === 'function') {
-    return ReflectApply(resolver, this, [hostname, callback]);
-  }
-  throw new ERR_INVALID_ARG_VALUE('rrtype', rrtype);
-}
-
 function defaultResolverSetServers(servers) {
   const resolver = new Resolver();
 
@@ -429,7 +340,7 @@ module.exports = {
   CANCELLED,
 };
 
-bindDefaultResolver(module.exports, getDefaultResolver());
+bindDefaultResolver(module.exports, Resolver.prototype);
 
 ObjectDefineProperties(module.exports, {
   promises: {
