+++ lib/internal/dns/promises.js
@@ -1,7 +1,6 @@
 'use strict';
 const {
   ArrayPrototypeMap,
-  ObjectCreate,
   ObjectDefineProperty,
   Promise,
   ReflectApply,
@@ -10,16 +9,15 @@ const {
 
 const {
   bindDefaultResolver,
-  Resolver: CallbackResolver,
+  createResolverClass,
   validateHints,
-  validateTimeout,
-  validateTries,
   emitInvalidHostnameWarning,
   getDefaultVerbatim,
   errorCodes: dnsErrorCodes,
   setDefaultResultOrder,
   setDefaultResolver,
 } = require('internal/dns/utils');
+
 const {
   NODATA,
   FORMERR,
@@ -52,7 +50,6 @@ const { isIP } = require('internal/net');
 const {
   getaddrinfo,
   getnameinfo,
-  ChannelWrap,
   GetAddrInfoReqWrap,
   GetNameInfoReqWrap,
   QueryReqWrap
@@ -305,36 +302,7 @@ function resolver(bindingName) {
   return query;
 }
 
-
-const resolveMap = ObjectCreate(null);
-
-// Resolver instances correspond 1:1 to c-ares channels.
-class Resolver {
-  constructor(options = undefined) {
-    const timeout = validateTimeout(options);
-    const tries = validateTries(options);
-    this._handle = new ChannelWrap(timeout, tries);
-  }
-}
-
-Resolver.prototype.getServers = CallbackResolver.prototype.getServers;
-Resolver.prototype.setServers = CallbackResolver.prototype.setServers;
-Resolver.prototype.cancel = CallbackResolver.prototype.cancel;
-Resolver.prototype.setLocalAddress = CallbackResolver.prototype.setLocalAddress;
-Resolver.prototype.resolveAny = resolveMap.ANY = resolver('queryAny');
-Resolver.prototype.resolve4 = resolveMap.A = resolver('queryA');
-Resolver.prototype.resolve6 = resolveMap.AAAA = resolver('queryAaaa');
-Resolver.prototype.resolveCaa = resolveMap.CAA = resolver('queryCaa');
-Resolver.prototype.resolveCname = resolveMap.CNAME = resolver('queryCname');
-Resolver.prototype.resolveMx = resolveMap.MX = resolver('queryMx');
-Resolver.prototype.resolveNs = resolveMap.NS = resolver('queryNs');
-Resolver.prototype.resolveTxt = resolveMap.TXT = resolver('queryTxt');
-Resolver.prototype.resolveSrv = resolveMap.SRV = resolver('querySrv');
-Resolver.prototype.resolvePtr = resolveMap.PTR = resolver('queryPtr');
-Resolver.prototype.resolveNaptr = resolveMap.NAPTR = resolver('queryNaptr');
-Resolver.prototype.resolveSoa = resolveMap.SOA = resolver('querySoa');
-Resolver.prototype.reverse = resolver('getHostByAddr');
-Resolver.prototype.resolve = function resolve(hostname, rrtype) {
+function resolve(hostname, rrtype) {
   let resolver;
 
   if (rrtype !== undefined) {
@@ -349,7 +317,11 @@ Resolver.prototype.resolve = function resolve(hostname, rrtype) {
   }
 
   return ReflectApply(resolver, this, [hostname]);
-};
+}
+
+// Promise-based resolver.
+const { Resolver, resolveMap } = createResolverClass(resolver);
+Resolver.prototype.resolve = resolve;
 
 function defaultResolverSetServers(servers) {
   const resolver = new Resolver();
