+++ lib/internal/dns/utils.js
@@ -9,6 +9,7 @@ const {
   NumberParseInt,
   RegExpPrototypeExec,
   RegExpPrototypeSymbolReplace,
+  ObjectCreate,
 } = primordials;
 
 const errors = require('internal/errors');
@@ -20,13 +21,11 @@ const {
   validateOneOf,
   validateString,
 } = require('internal/validators');
-const {
-  ChannelWrap,
-  strerror,
-  AI_ADDRCONFIG,
-  AI_ALL,
-  AI_V4MAPPED,
-} = internalBinding('cares_wrap');
+let binding;
+function lazyBinding() {
+  binding ??= internalBinding('cares_wrap');
+  return binding;
+}
 const IANA_DNS_PORT = 53;
 const IPv6RE = /^\[([^[\]]*)\]/;
 const addrSplitRE = /(^.+?)(?::(\d+))?$/;
@@ -49,10 +48,12 @@ function validateTries(options) {
 }
 
 // Resolver instances correspond 1:1 to c-ares channels.
-class Resolver {
+
+class ResolverBase {
   constructor(options = undefined) {
     const timeout = validateTimeout(options);
     const tries = validateTries(options);
+    const { ChannelWrap } = lazyBinding();
     this._handle = new ChannelWrap(timeout, tries);
   }
 
@@ -122,6 +123,7 @@ class Resolver {
     if (errorNumber !== 0) {
       // Reset the servers to the old servers, because ares probably unset them.
       this._handle.setServers(ArrayPrototypeJoin(orig, ','));
+      const { strerror } = lazyBinding();
       const err = strerror(errorNumber);
       throw new ERR_DNS_SET_SERVERS_FAILED(err, servers);
     }
@@ -138,7 +140,19 @@ class Resolver {
   }
 }
 
-let defaultResolver = new Resolver();
+let defaultResolver;
+let dnsOrder;
+
+function initializeDns() {
+  const orderFromCLI = getOptionValue('--dns-result-order');
+  if (!orderFromCLI) {
+    dnsOrder ??= 'verbatim';
+  } else {
+    // Allow the deserialized application to override order from CLI.
+    dnsOrder = orderFromCLI;
+  }
+}
+
 const resolverKeys = [
   'getServers',
   'resolve',
@@ -158,6 +172,11 @@ const resolverKeys = [
 ];
 
 function getDefaultResolver() {
+  // We do this here instead of pre-execution so that the default resolver is
+  // only ever created when the user loads any dns module.
+  if (defaultResolver === undefined) {
+    defaultResolver = new ResolverBase();
+  }
   return defaultResolver;
 }
 
@@ -166,12 +185,14 @@ function setDefaultResolver(resolver) {
 }
 
 function bindDefaultResolver(target, source) {
+  const defaultResolver = getDefaultResolver();
   ArrayPrototypeForEach(resolverKeys, (key) => {
     target[key] = FunctionPrototypeBind(source[key], defaultResolver);
   });
 }
 
 function validateHints(hints) {
+  const { AI_ADDRCONFIG, AI_ALL, AI_V4MAPPED } = lazyBinding();
   if ((hints & ~(AI_ADDRCONFIG | AI_ALL | AI_V4MAPPED)) !== 0) {
     throw new ERR_INVALID_ARG_VALUE('hints', hints);
   }
@@ -190,8 +211,6 @@ function emitInvalidHostnameWarning(hostname) {
   }
 }
 
-let dnsOrder = getOptionValue('--dns-result-order') || 'verbatim';
-
 function getDefaultVerbatim() {
   return dnsOrder !== 'ipv4first';
 }
@@ -199,6 +218,31 @@ function getDefaultVerbatim() {
 function setDefaultResultOrder(value) {
   validateOneOf(value, 'dnsOrder', ['verbatim', 'ipv4first']);
   dnsOrder = value;
+}
+
+function createResolverClass(resolver) {
+  const resolveMap = ObjectCreate(null);
+
+  class Resolver extends ResolverBase {}
+
+  Resolver.prototype.resolveAny = resolveMap.ANY = resolver('queryAny');
+  Resolver.prototype.resolve4 = resolveMap.A = resolver('queryA');
+  Resolver.prototype.resolve6 = resolveMap.AAAA = resolver('queryAaaa');
+  Resolver.prototype.resolveCaa = resolveMap.CAA = resolver('queryCaa');
+  Resolver.prototype.resolveCname = resolveMap.CNAME = resolver('queryCname');
+  Resolver.prototype.resolveMx = resolveMap.MX = resolver('queryMx');
+  Resolver.prototype.resolveNs = resolveMap.NS = resolver('queryNs');
+  Resolver.prototype.resolveTxt = resolveMap.TXT = resolver('queryTxt');
+  Resolver.prototype.resolveSrv = resolveMap.SRV = resolver('querySrv');
+  Resolver.prototype.resolvePtr = resolveMap.PTR = resolver('queryPtr');
+  Resolver.prototype.resolveNaptr = resolveMap.NAPTR = resolver('queryNaptr');
+  Resolver.prototype.resolveSoa = resolveMap.SOA = resolver('querySoa');
+  Resolver.prototype.reverse = resolver('getHostByAddr');
+
+  return {
+    resolveMap,
+    Resolver
+  };
 }
 
 // ERROR CODES
@@ -236,9 +280,10 @@ module.exports = {
   validateHints,
   validateTimeout,
   validateTries,
-  Resolver,
   emitInvalidHostnameWarning,
   getDefaultVerbatim,
   setDefaultResultOrder,
   errorCodes,
+  createResolverClass,
+  initializeDns
 };
