+++ lib/internal/process/pre_execution.js
@@ -89,6 +89,8 @@ function prepareMainThreadExecution(expandArgv1 = false,
   initializeDeprecations();
   initializeWASI();
 
+  require('internal/dns/utils').initializeDns();
+
   require('internal/v8/startup_snapshot').runDeserializeCallbacks();
 
   if (!initialzeModules) {
